@@ -1,6 +1,6 @@
 {
   "name": "@digibearapp/digibear-icons-vue",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "A free high-quality icon library for Vue 3.",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
@@ -18,39 +18,44 @@
     "lint": "vue-cli-service lint",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "generate-stories": "node ./bin/generate.js"
+    "generate-stories": "node ./bin/generate.mjs",
+    "update-git": "git add . && git commit -m $npm_package_version && git push"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@digibearapp/digibear-svg-icons": "^1.0.1",
+    "@babel/core": "^7.16.7",
+    "@digibearapp/digibear-svg-icons": "^1.0.5",
     "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.0.6",
-    "@storybook/addon-actions": "^6.3.12",
-    "@storybook/addon-essentials": "^6.3.12",
-    "@storybook/addon-links": "^6.3.12",
-    "@storybook/vue3": "^6.3.12",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
+    "@rollup/plugin-node-resolve": "^13.1.2",
+    "@storybook/addon-actions": "^6.4.9",
+    "@storybook/addon-essentials": "^6.4.9",
+    "@storybook/addon-links": "^6.4.9",
+    "@storybook/vue3": "^6.4.9",
+    "@typescript-eslint/eslint-plugin": "^5.8.1",
+    "@typescript-eslint/parser": "^5.8.1",
+    "@vue/cli-plugin-babel": "~4.5.15",
+    "@vue/cli-plugin-eslint": "~4.5.15",
+    "@vue/cli-plugin-typescript": "~4.5.15",
+    "@vue/cli-service": "~4.5.15",
+    "@vue/compiler-sfc": "^3.2.26",
+    "@vue/eslint-config-typescript": "^10.0.0",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.3",
-    "chalk": "^4.1.2",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.0.0",
-    "rollup": "^2.59.0",
-    "rollup-plugin-dts": "^4.0.0",
+    "chalk": "^5.0.0",
+    "eslint": "^8.6.0",
+    "eslint-plugin-vue": "^8.2.0",
+    "rollup": "^2.62.0",
+    "rollup-plugin-dts": "^4.1.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.30.0",
+    "rollup-plugin-typescript2": "^0.31.1",
     "rollup-plugin-vue": "^6.0.0",
-    "typescript": "~4.1.5",
+    "typescript": "~4.5.4",
     "vue": "^3.0.0",
-    "vue-loader": "^16.8.2"
+    "vue-loader": "^15.9.8"
+  },
+  "dependencies": {
+    "@digibearapp/digibear-common-types": "^1.0.6",
+    "@digibearapp/digibear-svg-core": "^1.0.5"
   },
   "peerDependencies": {
     "vue": "^3.0.0"
@@ -76,9 +81,5 @@
     "> 1%",
     "last 2 versions",
     "not dead"
-  ],
-  "dependencies": {
-    "@digibearapp/digibear-common-types": "^1.0.1",
-    "@digibearapp/digibear-svg-core": "^1.0.1"
-  }
+  ]
 }
