@@ -1,9 +1,7 @@
-module.exports = Object.freeze({
-    PREFIX:'db',
-    HEADER: `\
+export const PREFIX = "db";
+export const HEADER = `\
 // Code generated by Téodor Todorov (Creator of Digibear).
 // DO NOT EDIT.
 // Copyright 2021 Digibear.
 // All rights reserved.
-`,
-});
+`;
