@@ -955,10 +955,10 @@ import { dbActivity,
 		dbFaceSmile,
 		dbFaceSmileTeeth,
 		dbFaceSurprise,
-		dbFastForward,
 		dbFastForward10,
 		dbFastForward15,
 		dbFastForward5,
+		dbFastForward,
 		dbFileArrowDown,
 		dbFileArrowLeft,
 		dbFileArrowRight,
@@ -1016,16 +1016,16 @@ import { dbActivity,
 		dbGift,
 		dbGlasses3d,
 		dbGlasses,
+		dbGlobe1Line,
+		dbGlobe2Lines,
 		dbGlobeEast,
 		dbGlobeStand,
 		dbGlobeWest,
-		dbGlobe1Line,
-		dbGlobe2Lines,
 		dbGrapes,
-		dbGridDividers,
 		dbGrid2,
 		dbGrid2Plus,
 		dbGrid3,
+		dbGridDividers,
 		dbGrip,
 		dbHandFistBack,
 		dbHandFistFront,
@@ -1162,8 +1162,8 @@ import { dbActivity,
 		dbMedal,
 		dbMediaLibrary,
 		dbMegaphone,
-		dbMenuBurgeModern,
 		dbMenuBurger,
+		dbMenuBurgerModern,
 		dbMenuBurgerStairs,
 		dbMeteor,
 		dbMicrophone,
@@ -1299,13 +1299,13 @@ import { dbActivity,
 		dbRedoCompactRRect,
 		dbRedoCompactSmall,
 		dbRedoCompactTriangle,
-		dbRepeat,
 		dbRepeat1,
+		dbRepeat,
 		dbReplay,
-		dbRewind,
 		dbRewind10,
 		dbRewind15,
 		dbRewind5,
+		dbRewind,
 		dbRocket,
 		dbRocketTilt45,
 		dbRotateAndCrop,
@@ -1330,14 +1330,14 @@ import { dbActivity,
 		dbSkipBack,
 		dbSkipForward,
 		dbSleep,
-		dbSlider3Horizontal,
-		dbSlider3Vertical,
 		dbSliders2Horizontal,
 		dbSliders2Vertical,
+		dbSliders3Horizontal,
+		dbSliders3Vertical,
 		dbSmartphone,
 		dbSnowflake,
-		dbSortAlphaAz,
-		dbSortAlphaZa,
+		dbSortAlphaAZ,
+		dbSortAlphaZA,
 		dbSortNum19,
 		dbSortNum91,
 		dbSpeakerHigh,
@@ -2418,10 +2418,10 @@ export default {
 				"faceSmile",
 				"faceSmileTeeth",
 				"faceSurprise",
-				"fastForward",
 				"fastForward10",
 				"fastForward15",
 				"fastForward5",
+				"fastForward",
 				"fileArrowDown",
 				"fileArrowLeft",
 				"fileArrowRight",
@@ -2479,16 +2479,16 @@ export default {
 				"gift",
 				"glasses3d",
 				"glasses",
+				"globe1Line",
+				"globe2Lines",
 				"globeEast",
 				"globeStand",
 				"globeWest",
-				"globe1Line",
-				"globe2Lines",
 				"grapes",
-				"gridDividers",
 				"grid2",
 				"grid2Plus",
 				"grid3",
+				"gridDividers",
 				"grip",
 				"handFistBack",
 				"handFistFront",
@@ -2625,8 +2625,8 @@ export default {
 				"medal",
 				"mediaLibrary",
 				"megaphone",
-				"menuBurgeModern",
 				"menuBurger",
+				"menuBurgerModern",
 				"menuBurgerStairs",
 				"meteor",
 				"microphone",
@@ -2762,13 +2762,13 @@ export default {
 				"redoCompactRRect",
 				"redoCompactSmall",
 				"redoCompactTriangle",
-				"repeat",
 				"repeat1",
+				"repeat",
 				"replay",
-				"rewind",
 				"rewind10",
 				"rewind15",
 				"rewind5",
+				"rewind",
 				"rocket",
 				"rocketTilt45",
 				"rotateAndCrop",
@@ -2793,14 +2793,14 @@ export default {
 				"skipBack",
 				"skipForward",
 				"sleep",
-				"slider3Horizontal",
-				"slider3Vertical",
 				"sliders2Horizontal",
 				"sliders2Vertical",
+				"sliders3Horizontal",
+				"sliders3Vertical",
 				"smartphone",
 				"snowflake",
-				"sortAlphaAz",
-				"sortAlphaZa",
+				"sortAlphaAZ",
+				"sortAlphaZA",
 				"sortNum19",
 				"sortNum91",
 				"speakerHigh",
@@ -3893,10 +3893,10 @@ const Template: Story<DbIconProps> = (args, { argTypes }) => {
 		dbFaceSmile,
 		dbFaceSmileTeeth,
 		dbFaceSurprise,
-		dbFastForward,
 		dbFastForward10,
 		dbFastForward15,
 		dbFastForward5,
+		dbFastForward,
 		dbFileArrowDown,
 		dbFileArrowLeft,
 		dbFileArrowRight,
@@ -3954,16 +3954,16 @@ const Template: Story<DbIconProps> = (args, { argTypes }) => {
 		dbGift,
 		dbGlasses3d,
 		dbGlasses,
+		dbGlobe1Line,
+		dbGlobe2Lines,
 		dbGlobeEast,
 		dbGlobeStand,
 		dbGlobeWest,
-		dbGlobe1Line,
-		dbGlobe2Lines,
 		dbGrapes,
-		dbGridDividers,
 		dbGrid2,
 		dbGrid2Plus,
 		dbGrid3,
+		dbGridDividers,
 		dbGrip,
 		dbHandFistBack,
 		dbHandFistFront,
@@ -4100,8 +4100,8 @@ const Template: Story<DbIconProps> = (args, { argTypes }) => {
 		dbMedal,
 		dbMediaLibrary,
 		dbMegaphone,
-		dbMenuBurgeModern,
 		dbMenuBurger,
+		dbMenuBurgerModern,
 		dbMenuBurgerStairs,
 		dbMeteor,
 		dbMicrophone,
@@ -4237,13 +4237,13 @@ const Template: Story<DbIconProps> = (args, { argTypes }) => {
 		dbRedoCompactRRect,
 		dbRedoCompactSmall,
 		dbRedoCompactTriangle,
-		dbRepeat,
 		dbRepeat1,
+		dbRepeat,
 		dbReplay,
-		dbRewind,
 		dbRewind10,
 		dbRewind15,
 		dbRewind5,
+		dbRewind,
 		dbRocket,
 		dbRocketTilt45,
 		dbRotateAndCrop,
@@ -4268,14 +4268,14 @@ const Template: Story<DbIconProps> = (args, { argTypes }) => {
 		dbSkipBack,
 		dbSkipForward,
 		dbSleep,
-		dbSlider3Horizontal,
-		dbSlider3Vertical,
 		dbSliders2Horizontal,
 		dbSliders2Vertical,
+		dbSliders3Horizontal,
+		dbSliders3Vertical,
 		dbSmartphone,
 		dbSnowflake,
-		dbSortAlphaAz,
-		dbSortAlphaZa,
+		dbSortAlphaAZ,
+		dbSortAlphaZA,
 		dbSortNum19,
 		dbSortNum91,
 		dbSpeakerHigh,
